requirements.txt
pyparsing
appdirs
setuptools==40.1.0
cryptography==2.8
bcrypt==3.1.7
PyNaCl==1.3.0
Fabric3==1.14.post1
asgiref==3.7.2
attrs==22.2.0
cffi==1.15.1
ConfigParser==6.0.0
contextlib2==21.6.0
Cython==3.0.4
dl==0.1.0
docutils==0.20.1
filelock==3.12.4
gssapi==1.8.3
HTMLParser==0.0.2
importlib_metadata==6.8.0
invoke==2.2.0
ipython==8.16.1
ipywidgets==8.1.1
jnius==1.1.0
keyring==24.2.0
mock==5.1.0
nose==1.3.7
numpy==1.26.1
ordereddict==1.1
Pillow==10.0.1
Pillow==10.1.0
protobuf==4.24.4
pyasn1==0.5.0
pycryptodome==3.19.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
railroad==0.5.0
redis==5.0.1
Sphinx==7.2.6
SQLAlchemy==2.0.22
sspi==0.1.0
tornado==6.3.3
urllib3_secure_extra==0.1.0
watchdog==3.0.0
win32security==2.1.0
wincertstore==0.2.1
xmlrpclib==1.0.1